+++ .codecov.yml
@@ -1,16 +1,16 @@
 ignore:
-- "**/*.pb.go"
-- "**/*.pb.gw.go"
-- "**/*generated.go"
-- "**/*generated.deepcopy.go"
-- "**/*generated.openapi.go"
-- "**/*_test.go"
-- "pkg/client/.*"
-- "vendor/.*"
-- "test/.*"
-- "rust/**/error.rs"
-- "rust/numaflow-models/**" # ignore generated files
-- "rust/numaflow-grpc/**" # ignore generated files
+  - "**/*.pb.go"
+  - "**/*.pb.gw.go"
+  - "**/*generated.go"
+  - "**/*generated.deepcopy.go"
+  - "**/*generated.openapi.go"
+  - "**/*_test.go"
+  - "pkg/client/.*"
+  - "vendor/.*"
+  - "test/.*"
+  - "rust/**/error.rs"
+  - "rust/numaflow-models/**" # ignore generated files
+  - "rust/numaflow-pb/**" # ignore generated files
 coverage:
   status:
     patch: off
