+++ rust/Cargo.toml
@@ -4,7 +4,7 @@ workspace = { members = [
     "servesink",
     "serving",
     "numaflow-core",
-    "numaflow-grpc",
+    "numaflow-pb",
 ] }
 
 [[bin]]
@@ -23,6 +23,6 @@ backoff = { path = "backoff" }
 servesink = { path = "servesink" }
 serving = { path = "serving" }
 numaflow-core = { path = "numaflow-core" }
-numaflow-grpc = { path = "numaflow-grpc" }
+numaflow-pb = { path = "numaflow-pb" }
 tracing = "0.1.40"
 tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
