+++ rust/Makefile
@@ -3,7 +3,7 @@ SHELL:=/bin/bash
 .PHONY: generate
 generate:
 	$(MAKE) --directory numaflow-models generate
-	$(MAKE) --directory numaflow-grpc generate
+	$(MAKE) --directory numaflow-pb generate
 
 .PHONY: build
 build:
