+++ rust/numaflow-core/src/monovertex/forwarder.rs
@@ -538,9 +538,9 @@ mod tests {
     use chrono::Utc;
     use numaflow::source::{Message, Offset, SourceReadRequest};
     use numaflow::{sink, source, sourcetransform};
-    use numaflow_grpc::clients::sink::sink_client::SinkClient;
-    use numaflow_grpc::clients::source::source_client::SourceClient;
-    use numaflow_grpc::clients::sourcetransformer::source_transform_client::SourceTransformClient;
+    use numaflow_pb::clients::sink::sink_client::SinkClient;
+    use numaflow_pb::clients::source::source_client::SourceClient;
+    use numaflow_pb::clients::sourcetransformer::source_transform_client::SourceTransformClient;
     use tokio::sync::mpsc;
     use tokio::sync::mpsc::Sender;
     use tokio_util::sync::CancellationToken;
