+++ rust/numaflow-pb/src/clients.rs
