+++ rust/numaflow-pb/src/clients/mapstream.v1.rs
