+++ rust/numaflow-pb/src/clients/sessionreduce.v1.rs
